==> jupyter_notebooks/1_data_collection.ipynb <==
# **Data Collection Notebook**

## Objectives

* Fetch data from Kaggle and save as raw data
* Inspect the data and save it under outputs/datasets/collection

## Inputs

* Kaggle JSON file - authenticator token
* Kaggle data file - inputs/datasets/raw/house-prices/house-price/house_prices_records.csv
* Kaggle data file - inputs/datasets/raw/house-prices/house-price/inherited_house.csv

## Outputs

* Generate Dataset: outputs/datasets/collection/

---

# Change working directory

Accessing the current directory

import os
current_dir = os.getcwd()
current_dir

Making sure working in the child of the workspace directory

os.chdir('/workspaces/milestone-project-heritage-housing-issues')
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Install Python Packages

%pip install -r /workspaces/milestone-project-heritage-housing-issues/requirements.txt

---

# Fetch Data from Kaggle

* Install Kaggle

%pip install kaggle==1.5.12

* To recognise Kaggle token

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

* Defining the Kaggle dataset

KaggleDatasetPath = "codeinstitute/housing-prices-data"
DestinationFolder = "inputs/datasets/raw"
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

* Unzip the downloaded file, delete the zip file and delete the kaggle.json file

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Inspect the Kaggle Data

import pandas as pd
df = pd.read_csv(
    f"inputs/datasets/raw/house-price-20211124T154130Z-001/"
    f"house-price/house_prices_records.csv")
df.head()

* Dataframe summary

df.info()

---

# Push files to Repo

* If you do not need to push files to Repo, you may replace this section with "Conclusions and Next Steps" and state your conclusions and next steps.

import os
try:
    os.makedirs(name='outputs/datasets/collection')
except Exception as e:
    print(e)

df.to_csv(f"outputs/datasets/collection/HousePricesRecords.csv", index=False)
